Refactor css directory to styles
