+++ src/components/Card.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../css/Card.sass';
+import '../styles/Card.sass';
 import TagList from './Tag.js';
 
 function Card(props) {
