+++ src/components/Landing.js
@@ -1,4 +1,4 @@
-import '../css/Landing.sass'
+import '../styles/Landing.sass'
 import BankDemo from '../assets/landing-photos/bank-demo.svg';
 import About from './About.js';
 import Welcome from '../assets/landing-photos/welcome.jpg';
