+++ src/styles/ImposterInfo.sass
