+++ src/styles/Types.sass
@@ -37,12 +37,9 @@
   margin-left: auto
   margin-right: auto
 
-
-
 .imposter-content li 
   font-family: Source Sans Pro
   margin-left: 1em
-
 
 .imposter-content blockquote 
   font-family: Source Sans Pro
@@ -50,25 +47,14 @@
   font-style: italic
 
 /* Adjusts video player size based on screen size */
-@media (min-width: 600px) 
-
-
-
 @media (min-width: 768px) 
-
   .imposter-content .page-image
-      max-width: 40%
-  
-
-
+    max-width: 40%
 
 @media (min-width: 992px) 
 
   .imposter-content .float-video
-      float:right
-  
+    float:right
 
   .imposter-content .page-image
-      max-width: 30%
-  
-
+    max-width: 30%
